git-svn-id: http://lpm.org.ru/svn/kumir2@225 382dccb8-6cdf-11e0-9640-eb60159f4418

+++ src/plugins/kumircoderun/context.h
@@ -26,7 +26,7 @@ struct Context {
     QVector<Bytecode::Instruction> program;
     Bytecode::ElemType type;
     ContextRunMode runMode;
-    int moduleId;
+    quint8 moduleId;
     int algId;
     int lineNo;
 };
